add scripts to generate hundreds of prime numbers
The gen-dh-primes.sh is a basic attempt at parallelization with bash,
but it's not very efficient since it waits for a batch of N processes
(where N = number of CPUs) before launching the next batch, so some
CPUs are idle periodically.

The gen-dh-primes.py script uses Python's multiprocessing Pool to keep
all the CPUs busy until all the prime numbers are generated.